ADD convenience functions
It is desired to be able to store hashes in a convenient format
that contains all necessary data to recreate the hash
given the correct password is provided.

This provides also a starting point for a further
simplified interface where a user may *only* provide such a
string as well as an unhashed password.