ADD unified interface for all hashing modules
It is desired to have a unified interface for all ways of hashing a password.
The target is, that all modules provide an "easy-use" interface made up of:
- hashEncodePassword - Hashes a password and returns an encoded version of the hash
that encoded version can be used with "verifyPassword"
- verifyPassword - Checks whether a given password can be turned into a hash contained
in string that is the encoded version of the hash

Further it shall provide 2 "helper" procs if users wants to write their own stuff:
- hashPassword - Hashes the password, but does not encode it as hashEncodePassword does.
It only provides the base64 encoded hash as is
- encodeHash - Provides a way to turn a passwordHash and all data used to create the hash
into an encoded string version, taht can be used with "verifyPassword".

This commit implements this interface for the explicit algorithms:
- Argon2
- PBKDF2_HMAC_sha256
- PBKDF2_HMAC_sha512