REFACTOR encodeHash to have base64 encoded salts
In order to achieve similar behaviour as with argon2 from libsodium,
this library now stores salts in a base64 encoded format.

Further, to unify the interface, the salt for initial hashing
is now also generated via sysrand.

The benefit of this is that the user no longer needs to generate
the salt themselves.