REFACTOR Api regarding salt for consistency
Previously it was unclear if you had a string whether the salt you had
was actual just a byte-array cast to be a string,
or actually a base64 encoded string of said byte array.
To get rid of this confusion, salts now are byte-sequences everywhere.
The utils to help with this are in base64_utils.