Moved binding digestSizeProc to global space
DigestSizeProcs are procs bound from the openssl lib at runtime
and stored in the "sizeProc" variable.
It makes a lot more sense to do that in global space,
as that gets executed on startup of the application.
That way, if an error occurs, it occurs on startup
and not when the application is running.