#6 Make API more secure
It shouldn't be possible for passwords to accidentally leak.
Thus a new type was introduced: `Password`
Now all `hashPassword` procs only accept that type.
They also now return `seq[byte]` which is a more correct representation
of a hash.
In turn, encodeHash accepts seq[byte] and encodes it itself.

For clarity purposes the `Hash` type was introduced for this.